Moved all unit tests to new folder and out of source directory. Fixes #42
