+++ Source/Reference Design Libraries.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.539
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29025.244
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RFmxWLAN", "SA\RFmx WLAN\RFmxWLAN.csproj", "{9E8F4D28-AC1D-4A06-8C62-FC96097890B9}"
 EndProject
@@ -22,19 +22,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Modules", "Modules", "{9DF8
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Scope", "Scope\Scope.csproj", "{7659CA3A-45D7-4A1C-8253-98615F119CF3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScopeTests", "Unit Tests\ScopeTests\ScopeTests.csproj", "{62597F51-E95C-4303-B137-38D91CAE2B52}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Unit Tests", "Unit Tests", "{0C05E087-F135-408A-B150-64107CEF583B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Digital", "Digital\Digital.csproj", "{A12F87E8-772D-4DD9-9AE1-E6A72E83F4FE}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RFmxNR", "SA\RFmx NR\RFmxNR.csproj", "{B12280C1-69D7-4E52-8A41-95B1D568DD6B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SGTests", "Unit Tests\SGTests\SGTests.csproj", "{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RFmxLTE", "SA\RFmx LTE\RFmxLTE.csproj", "{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LTETests", "Unit Tests\LTETests\LTETests.csproj", "{2E7D656A-7813-436E-A217-61DEC8C2B907}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -62,9 +54,6 @@ Global
 		{7659CA3A-45D7-4A1C-8253-98615F119CF3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7659CA3A-45D7-4A1C-8253-98615F119CF3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7659CA3A-45D7-4A1C-8253-98615F119CF3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{62597F51-E95C-4303-B137-38D91CAE2B52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{62597F51-E95C-4303-B137-38D91CAE2B52}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{62597F51-E95C-4303-B137-38D91CAE2B52}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A12F87E8-772D-4DD9-9AE1-E6A72E83F4FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A12F87E8-772D-4DD9-9AE1-E6A72E83F4FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A12F87E8-772D-4DD9-9AE1-E6A72E83F4FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -73,17 +62,10 @@ Global
 		{B12280C1-69D7-4E52-8A41-95B1D568DD6B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B12280C1-69D7-4E52-8A41-95B1D568DD6B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B12280C1-69D7-4E52-8A41-95B1D568DD6B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2E7D656A-7813-436E-A217-61DEC8C2B907}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2E7D656A-7813-436E-A217-61DEC8C2B907}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2E7D656A-7813-436E-A217-61DEC8C2B907}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2E7D656A-7813-436E-A217-61DEC8C2B907}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -95,12 +77,9 @@ Global
 		{2B079B61-B71B-440E-BFF7-0DF80927B5C5} = {9DF8AA31-0EBB-4E3D-ACAE-96F2350DA9F8}
 		{3D1BC0F6-273C-4A44-82F6-3BB933862FFB} = {2B079B61-B71B-440E-BFF7-0DF80927B5C5}
 		{7659CA3A-45D7-4A1C-8253-98615F119CF3} = {9DF8AA31-0EBB-4E3D-ACAE-96F2350DA9F8}
-		{62597F51-E95C-4303-B137-38D91CAE2B52} = {0C05E087-F135-408A-B150-64107CEF583B}
 		{A12F87E8-772D-4DD9-9AE1-E6A72E83F4FE} = {9DF8AA31-0EBB-4E3D-ACAE-96F2350DA9F8}
 		{B12280C1-69D7-4E52-8A41-95B1D568DD6B} = {2B079B61-B71B-440E-BFF7-0DF80927B5C5}
-		{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5} = {0C05E087-F135-408A-B150-64107CEF583B}
 		{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236} = {2B079B61-B71B-440E-BFF7-0DF80927B5C5}
-		{2E7D656A-7813-436E-A217-61DEC8C2B907} = {0C05E087-F135-408A-B150-64107CEF583B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D020C05D-4EF8-45D5-8E12-D76DE5E66039}
