+++ build-tools/Unit Tests/RDLUnitTests.sln
@@ -1,0 +1,69 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29025.244
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LTETests", "LTETests\LTETests.csproj", "{2E7D656A-7813-436E-A217-61DEC8C2B907}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScopeTests", "ScopeTests\ScopeTests.csproj", "{62597F51-E95C-4303-B137-38D91CAE2B52}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SGTests", "SGTests\SGTests.csproj", "{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SG", "..\..\Source\SG\SG.csproj", "{89ABCAC5-CEDB-4535-BD58-E205A97834D9}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SG", "SG", "{B44BE09E-FE4C-4F6C-A17C-9C7C09308945}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Scope", "Scope", "{185DD0B8-D89C-41A0-8EF9-2C94DF1D3060}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "LTE", "LTE", "{6945D804-B83C-459D-99F6-DB89A02F8D77}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RFmxLTE", "..\..\Source\SA\RFmx LTE\RFmxLTE.csproj", "{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Scope", "..\..\Source\Scope\Scope.csproj", "{7659CA3A-45D7-4A1C-8253-98615F119CF3}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2E7D656A-7813-436E-A217-61DEC8C2B907}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2E7D656A-7813-436E-A217-61DEC8C2B907}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2E7D656A-7813-436E-A217-61DEC8C2B907}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2E7D656A-7813-436E-A217-61DEC8C2B907}.Release|Any CPU.Build.0 = Release|Any CPU
+		{62597F51-E95C-4303-B137-38D91CAE2B52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{62597F51-E95C-4303-B137-38D91CAE2B52}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{62597F51-E95C-4303-B137-38D91CAE2B52}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{62597F51-E95C-4303-B137-38D91CAE2B52}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{89ABCAC5-CEDB-4535-BD58-E205A97834D9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{89ABCAC5-CEDB-4535-BD58-E205A97834D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{89ABCAC5-CEDB-4535-BD58-E205A97834D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{89ABCAC5-CEDB-4535-BD58-E205A97834D9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7659CA3A-45D7-4A1C-8253-98615F119CF3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7659CA3A-45D7-4A1C-8253-98615F119CF3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7659CA3A-45D7-4A1C-8253-98615F119CF3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7659CA3A-45D7-4A1C-8253-98615F119CF3}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{2E7D656A-7813-436E-A217-61DEC8C2B907} = {6945D804-B83C-459D-99F6-DB89A02F8D77}
+		{62597F51-E95C-4303-B137-38D91CAE2B52} = {185DD0B8-D89C-41A0-8EF9-2C94DF1D3060}
+		{1A49ABF3-6F1E-4F3C-BCB6-513500F27DB5} = {B44BE09E-FE4C-4F6C-A17C-9C7C09308945}
+		{89ABCAC5-CEDB-4535-BD58-E205A97834D9} = {B44BE09E-FE4C-4F6C-A17C-9C7C09308945}
+		{1EB6F1AE-2EEA-4DC8-8CDB-09B5FADF2236} = {6945D804-B83C-459D-99F6-DB89A02F8D77}
+		{7659CA3A-45D7-4A1C-8253-98615F119CF3} = {185DD0B8-D89C-41A0-8EF9-2C94DF1D3060}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9E2F895C-40B9-4337-ACB4-6907CDBA042F}
+	EndGlobalSection
+EndGlobal
