+++ build-tools/Unit Tests/SGTests/Properties/AssemblyInfo.cs
