+++ build-tools/Unit Tests/SGTests/app.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+  <runtime>
+    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+      <dependentAssembly>
+        <assemblyIdentity name="NationalInstruments.Common" publicKeyToken="dc6ad606294fc298" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-19.0.40.49152" newVersion="19.0.40.49152" />
+      </dependentAssembly>
+    </assemblyBinding>
+  </runtime>
+</configuration>
