+++ build-tools/Unit Tests/ScopeTests/Properties/AssemblyInfo.cs
